independent CoLink class file
